+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Versions
 
 ## 2.29.1 (??)
+- Feature: take into account composed misspellings in the uppercase finder
 - Fix: escape some special characters before sending a regular expression to Wikipedia search
 - Performance: load the listings after the application has started to improve the deployment times in Toolforge